+++ lib/exshome/datatype/boolean.ex
@@ -1,0 +1,7 @@
+defmodule Exshome.DataType.Boolean do
+  @moduledoc """
+  Boolean datatype.
+  """
+
+  use Exshome.DataType, base_type: :boolean, icon: "✅", name: "boolean"
+end