+++ lib/exshome/datatype/integer.ex
@@ -1,0 +1,7 @@
+defmodule Exshome.DataType.Integer do
+  @moduledoc """
+  Integer datatype.
+  """
+
+  use Exshome.DataType, base_type: :integer, icon: "🔢", name: "integer"
+end